:hammer: Refactor:루트 레이아웃 header footer 컴포넌트 분리
